pyproject.toml
[project]
name = "file-reader"
version = "0.1.0"
description = "Hospital Patient Data Management MCP Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pandas>=2.0.0",
    "pyarrow>=20.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "joblib>=1.2.0",
    "numpy>=1.24.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "requests>=2.32.4",
    "langchain>=0.3.26",
    "streamlit>=1.46.1",
    "langchain-groq>=0.3.5",
    "langchain-openai>=0.3.0",
    "langgraph>=0.5.1",
    "langchain-mcp-adapters>=0.1.8",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "mcp-use>=1.3.6",
    "pypandoc>=1.15",
    "spacy>=3.8.7",
    "sentence-transformers>=5.0.0",
    "langdetect>=1.0.9",
    "deep-translator>=1.11.4",
    "nest-asyncio>=1.6.0",
    "shap>=0.44.0",
    "matplotlib>=3.8.0",
]

[project.scripts]
file-reader = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils", "tools"]
